Crate linera_alloy

Alloy

Alloy connects applications to blockchains.

Alloy is a rewrite of ethers-rs from the ground up, with exciting new features, high performance, and excellent docs.

ethers-rs will continue to be maintained until we have achieved feature-parity in Alloy. No action is currently needed from devs.

Installation

Currently, Alloy is not hosted on crates.io, the Rust package registry.

To incorporate Alloy into your project, you will need to specify the GitHub repository as the source. This can be achieved by executing the following command in your terminal:

cargo add alloy --git https://github.com/alloy-rs/alloy

After incorporating Alloy, you may wish to utilize specific features of the crate. These features can be enabled through modifications in your project’s Cargo.toml file. A comprehensive list of available features can be found at this GitHub link.

Overview

This repository contains the following crates:

alloy: Meta-crate for the entire project, including linera-alloy-core

linera-alloy-consensus - Ethereum consensus interface

linera-alloy-contract - Interact with on-chain contracts

linera-alloy-eips - Ethereum Improvement Proposal (EIP) implementations

linera-alloy-genesis - Ethereum genesis file definitions

linera-alloy-json-rpc - Core data types for JSON-RPC 2.0 clients

linera-alloy-network - Network abstraction for RPC types

linera-alloy-node-bindings - Ethereum execution-layer client bindings

linera-alloy-provider - Interface with an Ethereum blockchain

linera-alloy-pubsub - Ethereum JSON-RPC publish-subscribe tower service and type definitions

linera-alloy-rpc-client - Low-level Ethereum JSON-RPC client implementation

linera-alloy-rpc-types - Ethereum JSON-RPC types

linera-alloy-rpc-types-engine - Ethereum execution-consensus layer (engine) API RPC types

linera-alloy-rpc-types-trace - Ethereum RPC trace types

linera-alloy-signer - Ethereum signer abstraction

linera-alloy-signer-aws - AWS KMS signer implementation

linera-alloy-signer-gcp - GCP KMS signer implementation

linera-alloy-signer-ledger - Ledger signer implementation

linera-alloy-signer-trezor - Trezor signer implementation

linera-alloy-signer-wallet - Local wallet (Keystore/Mnemonic/Yubihsm) signer implementation

linera-alloy-transport - Low-level Ethereum JSON-RPC transport abstraction

linera-alloy-transport-http - HTTP transport implementation

linera-alloy-transport-ipc - IPC transport implementation

linera-alloy-transport-ws - WS transport implementation

Supported Rust Versions

Alloy will keep a rolling MSRV (minimum supported rust version) policy of at least 6 months. When increasing the MSRV, the new Rust version must have been released at least six months ago. The current MSRV is 1.76.

Note that the MSRV is not increased automatically, and only as part of a minor release.

Contributing

Thanks for your help improving the project! We are so happy to have you! We have a contributing guide to help you get involved in the Alloy project.

Pull requests will not be merged unless CI passes, so please ensure that your contribution follows the linting rules and passes clippy.

Note on no_std

Because these crates are primarily network-focused, we do not intend to support no_std for most of them at this time.

The following crates support no_std:

linera-alloy-eips

linera-alloy-genesis

linera-alloy-serde

linera-alloy-consensus

If you would like to add no_std support to a crate, please make sure to update scripts/check_no_std.sh as well.

Credits

None of these crates would have been possible without the great work done in:

ethers.js

rust-web3

ruint

ethabi

ethcontract-rs

guac_rs

License
Licensed under either of Apache License, Version 2.0 or MIT license at your option.
Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in these crates by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Re-exports

pub use primitives::hex;

pub use primitives::hex;

pub use primitives::uint;

pub use sol_types::sol;

Modules

contract

linera-alloy-contract

dyn_abi

linera-alloy-dyn-abi

json_abi

linera-alloy-json-abi

network

linera-alloy-networks

node_bindings

linera-alloy-node-bindings

primitives

linera-alloy-primitives

providers

Interface with an Ethereum blockchain.

rpc

Ethereum JSON-RPC client and types.

signers

Ethereum signer abstraction and implementations.

sol_types

Solidity type modeling and ABI and EIP-712 codec implementation.

transports

Low-level Ethereum JSON-RPC transport abstraction and implementations.