Struct linera_alloy::contract::EventPoller

pub struct EventPoller<T, E> {
    pub poller: PollerBuilder<T, (Uint<256, 4>,), Vec<Log>>,
    /* private fields */
}

An event poller.

Polling configuration is available through the poller field.

Fields
poller: PollerBuilder<T, (Uint<256, 4>,), Vec<Log>>

The inner poller.

Implementations

impl<T, E> EventPoller<T, E>
where T: Transport + Clone, E: SolEvent,

pub fn into_stream(self) -> impl Stream<Item = Result<(E, Log), Error>> + Unpin

Starts the poller and returns a stream that yields the decoded event and the raw log.

Note that this stream will not return None until the provider is dropped.