Module linera_alloy::network::eip2718

EIP-2718 traits.

Enums

Eip2718Error

EIP-2718 decoding errors.

Traits

Decodable2718

Decoding trait for EIP-2718 envelopes. These envelopes wrap a transaction or a receipt with a type flag.

Eip2718Envelope

An EIP-2718 envelope, blanket implemented for types that impl Encodable2718 and Decodable2718. This envelope is a wrapper around a transaction, or a receipt, or any other type that is differentiated by an EIP-2718 transaction type.

Encodable2718

Encoding trait for EIP-2718 envelopes. These envelopes wrap a transaction or a receipt with a type flag.