Struct linera_alloy::primitives::Bloom

#[repr(transparent)]
pub struct Bloom(pub FixedBytes<256>);

Ethereum 256 byte bloom filter.

Tuple Fields
0: FixedBytes<256>
Implementations

impl Bloom

pub const ZERO: Bloom = _

Array of Zero bytes.

pub const fn new(bytes: [u8; 256]) -> Bloom

Wraps the given byte array in this type.

pub const fn with_last_byte(x: u8) -> Bloom

Creates a new byte array with the last byte set to x.

pub const fn repeat_byte(byte: u8) -> Bloom

Creates a new byte array where all bytes are set to byte.

pub const fn len_bytes() -> usize

Returns the size of this array in bytes.

pub fn from_slice(src: &[u8]) -> Bloom

Create a new byte array from the given slice src.

For a fallible version, use the TryFrom<&[u8]> implementation.

Note

The given bytes are interpreted in big endian order.

Panics

If the length of src and the number of bytes in Self do not match.

pub fn left_padding_from(value: &[u8]) -> Bloom

Create a new byte array from the given slice src, left-padding it with zeroes if necessary.

Note

The given bytes are interpreted in big endian order.

Panics

Panics if src.len() > N.

pub fn right_padding_from(value: &[u8]) -> Bloom

Create a new byte array from the given slice src, right-padding it with zeroes if necessary.

Note

The given bytes are interpreted in big endian order.

Panics

Panics if src.len() > N.

pub const fn into_array(self) -> [u8; 256]

Returns the inner bytes array.

pub fn covers(&self, b: &Bloom) -> bool

Returns true if all bits set in b are also set in self.

pub const fn const_eq(&self, other: &Bloom) -> bool

Compile-time equality. NOT constant-time equality.

pub const fn bit_and(self, rhs: Bloom) -> Bloom

Computes the bitwise AND of two FixedBytes.

pub const fn bit_or(self, rhs: Bloom) -> Bloom

Computes the bitwise OR of two FixedBytes.

pub const fn bit_xor(self, rhs: Bloom) -> Bloom

Computes the bitwise XOR of two FixedBytes.

impl Bloom

pub const fn data(&self) -> &[u8; 256]

Returns a reference to the underlying data.

pub fn data_mut(&mut self) -> &mut [u8; 256]

Returns a mutable reference to the underlying data.

pub fn contains_input(&self, input: BloomInput<'_>) -> bool

Returns true if this bloom filter is a possible superset of the other bloom filter, admitting false positives.

Note: This method may return false positives. This is inherent to the bloom filter data structure.

pub const fn const_contains(self, other: Bloom) -> bool

Compile-time version of contains.

Note: This method may return false positives. This is inherent to the bloom filter data structure.

pub fn contains(&self, other: &Bloom) -> bool

Returns true if this bloom filter is a possible superset of the other bloom filter, admitting false positives.

Note: This method may return false positives. This is inherent to the bloom filter data structure.

pub fn accrue(&mut self, input: BloomInput<'_>)

Accrues the input into the bloom filter.

pub fn accrue_bloom(&mut self, bloom: &Bloom)

Accrues the input into the bloom filter.

pub fn m3_2048(&mut self, bytes: &[u8])

Specialised Bloom filter that sets three bits out of 2048, given an arbitrary byte sequence.

See Section 4.3.1 “Transaction Receipt” of the Ethereum Yellow Paper (page 6).

pub fn m3_2048_hashed(&mut self, hash: &FixedBytes<32>)

m3_2048 but with a pre-hashed input.

pub fn accrue_raw_log(&mut self, address: Address, topics: &[FixedBytes<32>])

Ingests a raw log into the bloom filter.

pub fn accrue_log(&mut self, log: &Log)

Ingests a log into the bloom filter.

pub fn contains_raw_log( &self, address: Address, topics: &[FixedBytes<32>], ) -> bool

True if the bloom filter contains a log with given address and topics.

Note: This method may return false positives. This is inherent to the bloom filter data structure.

pub fn contains_log(&self, log: &Log) -> bool

True if the bloom filter contains a log with given address and topics.

Note: This method may return false positives. This is inherent to the bloom filter data structure.