Struct linera_alloy::primitives::FixedBytes

#[repr(transparent)]
pub struct FixedBytes<const N: usize>(pub [u8; N]);

A byte array of fixed length ([u8; N]).

This type allows us to more tightly control serialization, deserialization. rlp encoding, decoding, and other type-level attributes for fixed-length byte arrays.

Users looking to prevent type-confusion between byte arrays of different lengths should use the wrap_fixed_bytes! macro to create a new fixed-length byte array type.

Tuple Fields
0: [u8; N]
Implementations

impl<const N: usize> FixedBytes<N>

pub const ZERO: FixedBytes<N> = _

Array of Zero bytes.

pub const fn new(bytes: [u8; N]) -> FixedBytes<N>

Wraps the given byte array in FixedBytes.

pub const fn with_last_byte(x: u8) -> FixedBytes<N>

Creates a new FixedBytes with the last byte set to x.

pub const fn repeat_byte(byte: u8) -> FixedBytes<N>

Creates a new FixedBytes where all bytes are set to byte.

pub const fn len_bytes() -> usize

Returns the size of this byte array (N).

pub const fn concat_const<const M: usize, const Z: usize>( self, other: FixedBytes<M>, ) -> FixedBytes<Z>

Concatenate two FixedBytes.

Due to constraints in the language, the user must specify the value of the output size Z.

Panics

Panics if Z is not equal to N + M.

pub fn from_slice(src: &[u8]) -> FixedBytes<N>

Create a new FixedBytes from the given slice src.

For a fallible version, use the TryFrom<&[u8]> implementation.

Note

The given bytes are interpreted in big endian order.

Panics

If the length of src and the number of bytes in Self do not match.

pub fn left_padding_from(value: &[u8]) -> FixedBytes<N>

Create a new FixedBytes from the given slice src, left-padding it with zeroes if necessary.

Note

The given bytes are interpreted in big endian order.

Panics

Panics if src.len() > N.

pub fn right_padding_from(value: &[u8]) -> FixedBytes<N>

Create a new FixedBytes from the given slice src, right-padding it with zeroes if necessary.

Note

The given bytes are interpreted in big endian order.

Panics

Panics if src.len() > N.

pub const fn as_slice(&self) -> &[u8]

Returns a slice containing the entire array. Equivalent to &s[..].

pub fn as_mut_slice(&mut self) -> &mut [u8]

Returns a mutable slice containing the entire array. Equivalent to &mut s[..].

pub fn covers(&self, other: &FixedBytes<N>) -> bool

Returns true if all bits set in self are also set in b.

pub const fn const_covers(self, other: FixedBytes<N>) -> bool

Returns true if all bits set in self are also set in b.

pub const fn const_eq(&self, other: &FixedBytes<N>) -> bool

Compile-time equality. NOT constant-time equality.

pub fn is_zero(&self) -> bool

Returns true if no bits are set.

pub const fn const_is_zero(&self) -> bool

Returns true if no bits are set.

pub const fn bit_and(self, rhs: FixedBytes<N>) -> FixedBytes<N>

Computes the bitwise AND of two FixedBytes.

pub const fn bit_or(self, rhs: FixedBytes<N>) -> FixedBytes<N>

Computes the bitwise OR of two FixedBytes.

pub const fn bit_xor(self, rhs: FixedBytes<N>) -> FixedBytes<N>

Computes the bitwise XOR of two FixedBytes.