Module linera_alloy::providers

Interface with an Ethereum blockchain.

See linera_alloy_provider for more details.

Modules

ext

Extended APIs for the provider module.

fillers

Transaction Fillers

layers

Useful layer implementations for the provider. Currently this module contains the AnvilLayer, AnvilProvider and ChainLayer types.

network

linera-alloy-networks

utils

Provider-related utilities.

Structs

EthCall

A builder for an "eth_call" request. This type is returned by the Provider::call method.

Identity

An identity layer that does nothing.

PendingTransaction

Represents a transaction that is yet to be confirmed a specified number of times.

PendingTransactionBuilder

A builder for configuring a pending transaction watcher.

PendingTransactionConfig

Configuration for watching a pending transaction.

ProviderBuilder

A builder for constructing a Provider from various layers.

RootProvider

The root provider manages the RPC client and the heartbeat. It is at the base of every provider stack.

RpcWithBlock

An RpcCall that takes an optional BlockId parameter. By default this will use “latest”.

Stack

A stack of two providers.

Traits

Network

Captures type info for network-specific RPC requests/responses.

Provider

Provider is parameterized with a network and a transport. The default transport is type-erased, but you can do Provider<Http, N>.

ProviderLayer

A layering abstraction in the vein of tower::Layer

WalletProvider

Trait for Providers, Fill stacks, etc, which contain NetworkSigner.

Type Aliases

FilterPollerBuilder

A task that polls the provider with eth_getFilterChanges, returning a list of R.

ReqwestProvider

Type alias for a RootProvider using the Http transport and a reqwest client.

TraceCallList

List of trace calls for use with Provider::trace_call_many