Trait linera_alloy::providers::ext::DebugApi

pub trait DebugApi<N, T>: Send + Sync {
    // Required methods
    fn debug_trace_transaction<'life0, 'async_trait>(
        &'life0 self,
        hash: FixedBytes<32>,
        trace_options: GethDebugTracingOptions,
    ) -> Pin<Box<dyn Future<Output = Result<GethTrace, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn debug_trace_block_by_hash<'life0, 'async_trait>(
        &'life0 self,
        block: FixedBytes<32>,
        trace_options: GethDebugTracingOptions,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<TraceResult<GethTrace, String>>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn debug_trace_block_by_number<'life0, 'async_trait>(
        &'life0 self,
        block: BlockNumberOrTag,
        trace_options: GethDebugTracingOptions,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<TraceResult<GethTrace, String>>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn debug_trace_call<'life0, 'async_trait>(
        &'life0 self,
        tx: TransactionRequest,
        block: BlockNumberOrTag,
        trace_options: GethDebugTracingCallOptions,
    ) -> Pin<Box<dyn Future<Output = Result<GethTrace, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn debug_trace_call_many<'life0, 'async_trait>(
        &'life0 self,
        txs: Vec<TransactionRequest>,
        block: BlockNumberOrTag,
        trace_options: GethDebugTracingCallOptions,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<GethTrace>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
}

Debug namespace rpc interface that gives access to several non-standard RPC methods.

Required Methods

fn debug_trace_transaction<'life0, 'async_trait>( &'life0 self, hash: FixedBytes<32>, trace_options: GethDebugTracingOptions, ) -> Pin<Box<dyn Future<Output = Result<GethTrace, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Reruns the transaction specified by the hash and returns the trace.

It will replay any prior transactions to achieve the same state the transaction was executed in.

GethDebugTracingOptions can be used to specify the trace options.

Note

Not all nodes support this call.

fn debug_trace_block_by_hash<'life0, 'async_trait>( &'life0 self, block: FixedBytes<32>, trace_options: GethDebugTracingOptions, ) -> Pin<Box<dyn Future<Output = Result<Vec<TraceResult<GethTrace, String>>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Return a full stack trace of all invoked opcodes of all transaction that were included in this block.

The parent of the block must be present or it will fail.

GethDebugTracingOptions can be used to specify the trace options.

Note

Not all nodes support this call.

fn debug_trace_block_by_number<'life0, 'async_trait>( &'life0 self, block: BlockNumberOrTag, trace_options: GethDebugTracingOptions, ) -> Pin<Box<dyn Future<Output = Result<Vec<TraceResult<GethTrace, String>>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Same as debug_trace_block_by_hash but block is specified by number.

GethDebugTracingOptions can be used to specify the trace options.

Note

Not all nodes support this call.

fn debug_trace_call<'life0, 'async_trait>( &'life0 self, tx: TransactionRequest, block: BlockNumberOrTag, trace_options: GethDebugTracingCallOptions, ) -> Pin<Box<dyn Future<Output = Result<GethTrace, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Executes the given transaction without publishing it like eth_call and returns the trace of the execution.

The transaction will be executed in the context of the given block number or tag. The state its run on is the state of the previous block.

GethDebugTracingOptions can be used to specify the trace options.

Note

Not all nodes support this call.

fn debug_trace_call_many<'life0, 'async_trait>( &'life0 self, txs: Vec<TransactionRequest>, block: BlockNumberOrTag, trace_options: GethDebugTracingCallOptions, ) -> Pin<Box<dyn Future<Output = Result<Vec<GethTrace>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Same as debug_trace_call but it used to run and trace multiple transactions at once.

GethDebugTracingOptions can be used to specify the trace options.

Note

Not all nodes support this call.