Trait linera_alloy::providers::network::eip2718::Encodable2718

pub trait Encodable2718: Sized + Send + Sync + 'static {
    // Required methods
    fn type_flag(&self) -> Option<u8>;
    fn encode_2718_len(&self) -> usize;
    fn encode_2718(&self, out: &mut dyn BufMut);

    // Provided methods
    fn is_legacy(&self) -> bool { ... }
    fn encoded_2718(&self) -> Vec<u8>  { ... }
    fn trie_hash(&self) -> FixedBytes<32> { ... }
    fn seal(self) -> Sealed<Self> { ... }
    fn network_encode(&self, out: &mut dyn BufMut) { ... }
}

Encoding trait for EIP-2718 envelopes. These envelopes wrap a transaction or a receipt with a type flag.

Required Methods

fn type_flag(&self) -> Option<u8>

Return the type flag (if any).

This should return None for the default (legacy) variant of the envelope.

fn encode_2718_len(&self) -> usize

The length of the 2718 encoded envelope. This is the length of the type flag + the length of the inner transaction RLP.

fn encode_2718(&self, out: &mut dyn BufMut)

Encode the transaction according to EIP-2718 rules. First a 1-byte type flag in the range 0x0-0x7f, then the body of the transaction.

This implementation uses RLP for the transaction body. Non-standard users can override this to use some other serialization scheme.

Provided Methods

fn is_legacy(&self) -> bool

True if the envelope is the legacy variant.

fn encoded_2718(&self) -> Vec<u8>

Encode the transaction according to [EIP-2718] rules. First a 1-byte type flag in the range 0x0-0x7f, then the body of the transaction.

This is a convenience method for encoding into a vec, and returning the vec.

fn trie_hash(&self) -> FixedBytes<32>

Compute the hash as committed to in the MPT trie.

fn seal(self) -> Sealed<Self>

Seal the encodable, by encoding and hashing it.

fn network_encode(&self, out: &mut dyn BufMut)

Return the network encoding. For non-legacy items, this is the RLP encoding of the bytestring of the 2718 encoding. For legacy items it is simply the legacy encoding.

Object Safety

This trait is not object safe.