Struct linera_alloy::providers::RootProvider

pub struct RootProvider<T, N = Ethereum> { /* private fields */ }

The root provider manages the RPC client and the heartbeat. It is at the base of every provider stack.

Implementations

impl<N> RootProvider<Http<Client>, N>
where N: Network,

pub fn new_http(url: Url) -> RootProvider<Http<Client>, N>

Creates a new HTTP root provider from the given URL.

impl<T, N> RootProvider<T, N>
where T: Transport, N: Network,

pub fn new(client: RpcClient<T>) -> RootProvider<T, N>

Creates a new root provider from the given RPC client.

impl<N> RootProvider<BoxTransport, N>
where N: Network,

pub async fn connect_boxed<C>( conn: C, ) -> Result<RootProvider<BoxTransport, N>, RpcError<TransportErrorKind>>
where C: BoxTransportConnect,

Connects to a boxed transport with the given connector.

pub async fn connect_builtin( s: &str, ) -> Result<RootProvider<BoxTransport, N>, RpcError<TransportErrorKind>>

Creates a new root provider from the provided connection details.

impl<T, N> RootProvider<T, N>
where T: Transport + Clone, N: Network,

pub fn boxed(self) -> RootProvider<BoxTransport, N>

Boxes the inner client.

This will create a new provider if this instance is not the only reference to the inner client.