Enum linera_alloy::rpc::types::eth::FilterBlockOption

pub enum FilterBlockOption {
    Range {
        from_block: Option<BlockNumberOrTag>,
        to_block: Option<BlockNumberOrTag>,
    },
    AtBlockHash(FixedBytes<32>),
}

Represents the target range of blocks for the filter

Variants

Range

Represents a range of blocks with optional from and to blocks

Note: ranges are considered to be inclusive

Fields

from_block: Option<BlockNumberOrTag>

The block number or tag this filter should start at.

to_block: Option<BlockNumberOrTag>

The block number or that this filter should end at.

AtBlockHash(FixedBytes<32>)

The hash of the block if the filter only targets a single block

Implementations

impl FilterBlockOption

pub const fn get_to_block(&self) -> Option<&BlockNumberOrTag>

Returns the fromBlock value, if any

pub const fn get_from_block(&self) -> Option<&BlockNumberOrTag>

Returns the toBlock value, if any

pub const fn as_range( &self, ) -> (Option<&BlockNumberOrTag>, Option<&BlockNumberOrTag>)

Returns the range (fromBlock, toBlock) if this is a range filter.

impl FilterBlockOption

pub const fn set_from_block(&self, block: BlockNumberOrTag) -> FilterBlockOption

Sets the block number this range filter should start at.

pub const fn set_to_block(&self, block: BlockNumberOrTag) -> FilterBlockOption

Sets the block number this range filter should end at.

pub const fn set_hash(&self, hash: FixedBytes<32>) -> FilterBlockOption

Pins the block hash this filter should target.