Struct linera_alloy::rpc::types::eth::Log

pub struct Log<T = LogData> {
    pub inner: Log<T>,
    pub block_hash: Option<FixedBytes<32>>,
    pub block_number: Option<u64>,
    pub block_timestamp: Option<u64>,
    pub transaction_hash: Option<FixedBytes<32>>,
    pub transaction_index: Option<u64>,
    pub log_index: Option<u64>,
    pub removed: bool,
}

Ethereum Log emitted by a transaction

Fields
inner: Log<T>

Consensus log object

block_hash: Option<FixedBytes<32>>

Hash of the block the transaction that emitted this log was mined in

block_number: Option<u64>

Number of the block the transaction that emitted this log was mined in

block_timestamp: Option<u64>

The timestamp of the block as proposed in: https://ethereum-magicians.org/t/proposal-for-adding-blocktimestamp-to-logs-object-returned-by-eth-getlogs-and-related-requests https://github.com/ethereum/execution-apis/issues/295

transaction_hash: Option<FixedBytes<32>>

Transaction Hash

transaction_index: Option<u64>

Index of the Transaction in the block

log_index: Option<u64>

Log Index in Block

removed: bool

Geth Compatibility Field: whether this log was removed

Implementations

impl<T> Log<T>

pub const fn address(&self) -> Address

Getter for the address field. Shortcut for log.inner.address.

pub const fn data(&self) -> &T

Getter for the data field. Shortcut for log.inner.data.

impl Log

pub fn topics(&self) -> &[FixedBytes<32>]

Getter for the topics field. Shortcut for log.inner.topics().

pub fn topic0(&self) -> Option<&FixedBytes<32>>

Getter for the topic0 field.

pub fn topics_mut(&mut self) -> &mut [FixedBytes<32>]

Get the topic list, mutably. This gives access to the internal array, without allowing extension of that array. Shortcut for LogData::topics_mut

pub fn log_decode<T>(&self) -> Result<Log<T>, Error>
where T: SolEvent,

Decode the log data into a typed log.

impl<T> Log<T>
where &'a T: for<'a> Into<LogData>,

pub fn reserialize_inner(&self) -> Log

Reserialize the inner data, returning an linera_alloy_primitives::Log.

pub fn reserialize(&self) -> Log

Reserialize the data, returning a new Log object wrapping an linera_alloy_primitives::Log. this copies the log metadata, preserving the original object.