Struct linera_alloy::rpc::types::eth::txpool::TxpoolStatus

pub struct TxpoolStatus {
    pub pending: u64,
    pub queued: u64,
}

Transaction Pool Status

The status inspection property can be queried for the number of transactions currently pending for inclusion in the next block(s), as well as the ones that are being scheduled for future execution only.

See here for more details

Fields
pending: u64

number of pending tx

queued: u64

number of queued tx