Crate linera_alloy::signers::k256::ecdsa::signature::digest::block_buffer

Fixed size buffer for block processing of data.

Modules

generic_array

This crate implements a structure that can be used as a generic array type. Core Rust array types [T; N] can’t be used generically with respect to N, so for example this:

Structs

BlockBuffer

Buffer for block processing of data.

Eager

Eager block buffer kind, which guarantees that buffer position always lies in the range of 0..BlockSize.

Error

Block buffer error.

Lazy

Lazy block buffer kind, which guarantees that buffer position always lies in the range of 0..=BlockSize.

Traits

BufferKind

Trait for buffer kinds.

Type Aliases

Block

Block on which BlockBuffer operates.

EagerBuffer

Eager block buffer.

LazyBuffer

Lazy block buffer.