Trait linera_alloy::signers::k256::ecdsa::signature::digest::block_buffer::generic_array::functional::FunctionalSequence

pub unsafe trait FunctionalSequence<T>: GenericSequence<T> {
    // Provided methods
    fn map<U, F>(self, f: F) -> <Self::Mapped as GenericSequence<U>>::Sequence
       where Self: MappedGenericSequence<T, U>,
             Self::Length: ArrayLength<U>,
             F: FnMut(Self::Item) -> U { ... }
    fn zip<B, Rhs, U, F>(
        self,
        rhs: Rhs,
        f: F,
    ) -> <Self::Mapped as GenericSequence<U>>::Sequence
       where Self: MappedGenericSequence<T, U>,
             Rhs: MappedGenericSequence<B, U, Mapped = <Self::Mapped as GenericSequence<U>>::Sequence, Length = Self::Length> + GenericSequence<B>,
             Self::Length: ArrayLength<B> + ArrayLength<U>,
             F: FnMut(Self::Item, <Rhs as IntoIterator>::Item) -> U { ... }
    fn fold<U, F>(self, init: U, f: F) -> U
       where F: FnMut(U, Self::Item) -> U { ... }
}

Defines functional programming methods for generic sequences

Provided Methods

fn map<U, F>(self, f: F) -> <Self::Mapped as GenericSequence<U>>::Sequence
where Self: MappedGenericSequence<T, U>, Self::Length: ArrayLength<U>, F: FnMut(Self::Item) -> U,

Maps a GenericSequence to another GenericSequence.

If the mapping function panics, any already initialized elements in the new sequence will be dropped, AND any unused elements in the source sequence will also be dropped.

fn zip<B, Rhs, U, F>( self, rhs: Rhs, f: F, ) -> <Self::Mapped as GenericSequence<U>>::Sequence
where Self: MappedGenericSequence<T, U>, Rhs: MappedGenericSequence<B, U, Mapped = <Self::Mapped as GenericSequence<U>>::Sequence, Length = Self::Length> + GenericSequence<B>, Self::Length: ArrayLength<B> + ArrayLength<U>, F: FnMut(Self::Item, <Rhs as IntoIterator>::Item) -> U,

Combines two GenericSequence instances and iterates through both of them, initializing a new GenericSequence with the result of the zipped mapping function.

If the mapping function panics, any already initialized elements in the new sequence will be dropped, AND any unused elements in the source sequences will also be dropped.

fn fold<U, F>(self, init: U, f: F) -> U
where F: FnMut(U, Self::Item) -> U,

Folds (or reduces) a sequence of data into a single value.

If the fold function panics, any unused elements will be dropped.

Object Safety

This trait is not object safe.