Struct linera_alloy::signers::k256::ecdsa::signature::digest::block_buffer::generic_array::typenum::TArr

pub struct TArr<V, A> { /* private fields */ }

TArr is a type that acts as an array of types. It is defined similarly to UInt, only its values can be more than bits, and it is designed to act as an array. So you can only add two if they have the same number of elements, for example.

This array is only really designed to contain Integer types. If you use it with others, you may find it lacking functionality.