Struct linera_alloy::signers::k256::ecdsa::signature::digest::block_buffer::generic_array::typenum::UInt

pub struct UInt<U, B> { /* private fields */ }

UInt is defined recursively, where B is the least significant bit and U is the rest of the number. Conceptually, U should be bound by the trait Unsigned and B should be bound by the trait Bit, but enforcing these bounds causes linear instead of logrithmic scaling in some places, so they are left off for now. They may be enforced in future.

In order to keep numbers unique, leading zeros are not allowed, so UInt<UTerm, B0> is forbidden.

Example

use typenum::{UInt, UTerm, B0, B1};

type U6 = UInt<UInt<UInt<UTerm, B1>, B1>, B0>;

Implementations

impl<U, B> UInt<U, B>
where U: Unsigned, B: Bit,

pub fn new() -> UInt<U, B>

Instantiates a singleton representing this unsigned integer.