Trait linera_alloy::signers::k256::ecdsa::signature::digest::core_api::ExtendableOutputCore

pub trait ExtendableOutputCore: UpdateCore + BufferKindUser
where
    Self::BlockSize: IsLess<UInt<UInt<UInt<UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>, B0>, B0>, B0>>,
    <Self::BlockSize as IsLess<UInt<UInt<UInt<UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>, B0>, B0>, B0>>>::Output: NonZero,
{
    type ReaderCore: XofReaderCore;

    // Required method
    fn finalize_xof_core(
        &mut self,
        buffer: &mut BlockBuffer<Self::BlockSize, Self::BufferKind>,
    ) -> Self::ReaderCore;
}

Core trait for hash functions with extendable (XOF) output size.

Required Associated Types

type ReaderCore: XofReaderCore

XOF reader core state.

Required Methods

fn finalize_xof_core( &mut self, buffer: &mut BlockBuffer<Self::BlockSize, Self::BufferKind>, ) -> Self::ReaderCore

Retrieve XOF reader using remaining data stored in the block buffer and leave hasher in a dirty state.

Object Safety

This trait is not object safe.