Trait linera_alloy::signers::k256::ecdsa::signature::digest::crypto_common::KeyIvInit

pub trait KeyIvInit: Sized + KeySizeUser + IvSizeUser {
    // Required method
    fn new(
        key: &GenericArray<u8, Self::KeySize>,
        iv: &GenericArray<u8, Self::IvSize>,
    ) -> Self;

    // Provided method
    fn new_from_slices(key: &[u8], iv: &[u8]) -> Result<Self, InvalidLength> { ... }
}

Types which can be initialized from key and initialization vector (nonce).

Required Methods

fn new( key: &GenericArray<u8, Self::KeySize>, iv: &GenericArray<u8, Self::IvSize>, ) -> Self

Create new value from fixed length key and nonce.

Provided Methods

fn new_from_slices(key: &[u8], iv: &[u8]) -> Result<Self, InvalidLength>

Create new value from variable length key and nonce.

Object Safety

This trait is not object safe.