Trait linera_alloy::signers::k256::ecdsa::signature::RandomizedDigestSigner

pub trait RandomizedDigestSigner<D, S>
where
    D: Digest,
{
    // Required method
    fn try_sign_digest_with_rng(
        &self,
        rng: &mut impl CryptoRngCore,
        digest: D,
    ) -> Result<S, Error>;

    // Provided method
    fn sign_digest_with_rng(&self, rng: &mut impl CryptoRngCore, digest: D) -> S { ... }
}

Combination of DigestSigner and RandomizedSigner with support for computing a signature over a digest which requires entropy from an RNG.

Required Methods

fn try_sign_digest_with_rng( &self, rng: &mut impl CryptoRngCore, digest: D, ) -> Result<S, Error>

Attempt to sign the given prehashed message Digest, returning a digital signature on success, or an error if something went wrong.

Provided Methods

fn sign_digest_with_rng(&self, rng: &mut impl CryptoRngCore, digest: D) -> S

Sign the given prehashed message Digest, returning a signature.

Panics in the event of a signing error.

Object Safety

This trait is not object safe.