Struct linera_alloy::signers::k256::elliptic_curve::bigint::modular::runtime_mod::DynResidue

pub struct DynResidue<const LIMBS: usize> { /* private fields */ }

A residue represented using LIMBS limbs. The odd modulus of this residue is set at runtime.

Implementations

impl<const LIMBS: usize> DynResidue<LIMBS>

pub const fn add(&self, rhs: &DynResidue<LIMBS>) -> DynResidue<LIMBS>

Adds rhs.

impl<const LIMBS: usize> DynResidue<LIMBS>

pub const fn invert(&self) -> (DynResidue<LIMBS>, CtChoice)

Computes the residue self^-1 representing the multiplicative inverse of self. I.e. self * self^-1 = 1. If the number was invertible, the second element of the tuple is the truthy value, otherwise it is the falsy value (in which case the first element’s value is unspecified).

impl<const LIMBS: usize> DynResidue<LIMBS>

pub const fn mul(&self, rhs: &DynResidue<LIMBS>) -> DynResidue<LIMBS>

Multiplies by rhs.

pub const fn square(&self) -> DynResidue<LIMBS>

Computes the (reduced) square of a residue.

impl<const LIMBS: usize> DynResidue<LIMBS>

pub const fn neg(&self) -> DynResidue<LIMBS>

Negates the number.

impl<const LIMBS: usize> DynResidue<LIMBS>

pub const fn pow<const RHS_LIMBS: usize>( &self, exponent: &Uint<RHS_LIMBS>, ) -> DynResidue<LIMBS>

Raises to the exponent power.

pub const fn pow_bounded_exp<const RHS_LIMBS: usize>( &self, exponent: &Uint<RHS_LIMBS>, exponent_bits: usize, ) -> DynResidue<LIMBS>

Raises to the exponent power, with exponent_bits representing the number of (least significant) bits to take into account for the exponent.

NOTE: exponent_bits may be leaked in the time pattern.

impl<const LIMBS: usize> DynResidue<LIMBS>

pub const fn sub(&self, rhs: &DynResidue<LIMBS>) -> DynResidue<LIMBS>

Subtracts rhs.

impl<const LIMBS: usize> DynResidue<LIMBS>

pub const fn new( integer: &Uint<LIMBS>, residue_params: DynResidueParams<LIMBS>, ) -> DynResidue<LIMBS>

Instantiates a new Residue that represents this integer mod MOD.

pub const fn retrieve(&self) -> Uint<LIMBS>

Retrieves the integer currently encoded in this Residue, guaranteed to be reduced.

pub const fn zero(residue_params: DynResidueParams<LIMBS>) -> DynResidue<LIMBS>

Instantiates a new Residue that represents zero.

pub const fn one(residue_params: DynResidueParams<LIMBS>) -> DynResidue<LIMBS>

Instantiates a new Residue that represents 1.

pub const fn params(&self) -> &DynResidueParams<LIMBS>

Returns the parameter struct used to initialize this residue.

pub const fn as_montgomery(&self) -> &Uint<LIMBS>

Access the DynResidue value in Montgomery form.

pub fn as_montgomery_mut(&mut self) -> &mut Uint<LIMBS>

Mutably access the DynResidue value in Montgomery form.

pub const fn from_montgomery( integer: Uint<LIMBS>, residue_params: DynResidueParams<LIMBS>, ) -> DynResidue<LIMBS>

Create a DynResidue from a value in Montgomery form.

pub const fn to_montgomery(&self) -> Uint<LIMBS>

Extract the value from the DynResidue in Montgomery form.

pub fn div_by_2(&self) -> DynResidue<LIMBS>

Performs the modular division by 2, that is for given x returns y such that y * 2 = x mod p. This means:

if x is even, returns x / 2,

if x is odd, returns (x + p) / 2 (since the modulus p in Montgomery form is always odd, this divides entirely).