Trait linera_alloy::signers::k256::elliptic_curve::bigint::prelude::ConcatMixed

pub trait ConcatMixed<Lo = Self>
where
    Lo: ?Sized,
{
    type MixedOutput;

    // Required method
    fn concat_mixed(&self, lo: &Lo) -> Self::MixedOutput;
}

Concatenate two numbers into a “wide” combined-width value, using the lo value as the least significant value.

Required Associated Types

type MixedOutput

Concatenated output: combination of Lo and Self.

Required Methods

fn concat_mixed(&self, lo: &Lo) -> Self::MixedOutput

Concatenate the two values, with self as most significant and lo as the least significant.