Struct linera_alloy::signers::k256::elliptic_curve::bigint::Checked

pub struct Checked<T>(pub CtOption<T>);

Provides intentionally-checked arithmetic on T.

Internally this leverages the CtOption type from the subtle crate in order to handle overflows.

Tuple Fields
0: CtOption<T>
Implementations

impl<T> Checked<T>

pub fn new(val: T) -> Checked<T>

Create a new checked arithmetic wrapper for the given value.