Struct linera_alloy::signers::k256::elliptic_curve::bigint::Limb

#[repr(transparent)]
pub struct Limb(pub u64);

Big integers are represented as an array of smaller CPU word-size integers called “limbs”.

Tuple Fields
0: u64
Implementations

impl Limb

pub const fn adc(self, rhs: Limb, carry: Limb) -> (Limb, Limb)

Computes self + rhs + carry, returning the result along with the new carry.

pub const fn saturating_add(&self, rhs: Limb) -> Limb

Perform saturating addition.

pub const fn wrapping_add(&self, rhs: Limb) -> Limb

Perform wrapping addition, discarding overflow.

impl Limb

pub const fn bitand(self, rhs: Limb) -> Limb

Calculates a & b.

impl Limb

pub const fn not(self) -> Limb

Calculates !a.

impl Limb

pub const fn bitor(self, rhs: Limb) -> Limb

Calculates a | b.

impl Limb

pub const fn bitxor(self, rhs: Limb) -> Limb

Calculates a ^ b.

impl Limb

pub const fn bits(self) -> usize

Calculate the number of bits needed to represent this number.

pub const fn leading_zeros(self) -> usize

Calculate the number of leading zeros in the binary representation of this number.

pub const fn trailing_zeros(self) -> usize

Calculate the number of trailing zeros in the binary representation of this number.

pub const fn trailing_ones(self) -> usize

Calculate the number of trailing ones the binary representation of this number.

impl Limb

pub fn is_odd(&self) -> Choice

Is this limb an odd number?

pub fn cmp_vartime(&self, other: &Limb) -> Ordering

Perform a comparison of the inner value in variable-time.

Note that the PartialOrd and Ord impls wrap constant-time comparisons using the subtle crate.

pub const fn eq_vartime(&self, other: &Limb) -> bool

Performs an equality check in variable-time.

impl Limb

pub const fn from_u8(n: u8) -> Limb

Create a Limb from a u8 integer (const-friendly)

pub const fn from_u16(n: u16) -> Limb

Create a Limb from a u16 integer (const-friendly)

pub const fn from_u32(n: u32) -> Limb

Create a Limb from a u32 integer (const-friendly)

pub const fn from_u64(n: u64) -> Limb

Create a Limb from a u64 integer (const-friendly)

impl Limb

pub const fn mac(self, b: Limb, c: Limb, carry: Limb) -> (Limb, Limb)

Computes self + (b * c) + carry, returning the result along with the new carry.

pub const fn saturating_mul(&self, rhs: Limb) -> Limb

Perform saturating multiplication.

pub const fn wrapping_mul(&self, rhs: Limb) -> Limb

Perform wrapping multiplication, discarding overflow.

impl Limb

pub const fn wrapping_neg(self) -> Limb

Perform wrapping negation.

impl Limb

pub const fn shl(self, rhs: Limb) -> Limb

Computes self << rhs. Panics if rhs overflows Limb::BITS.

impl Limb

pub const fn shr(self, rhs: Limb) -> Limb

Computes self >> rhs. Panics if rhs overflows Limb::BITS.

impl Limb

pub const fn sbb(self, rhs: Limb, borrow: Limb) -> (Limb, Limb)

Computes self - (rhs + borrow), returning the result along with the new borrow.

pub const fn saturating_sub(&self, rhs: Limb) -> Limb

Perform saturating subtraction.

pub const fn wrapping_sub(&self, rhs: Limb) -> Limb

Perform wrapping subtraction, discarding underflow and wrapping around the boundary of the type.

impl Limb

pub const ZERO: Limb = _

The value 0.

pub const ONE: Limb = _

The value 1.

pub const MAX: Limb = _

Maximum value this Limb can express.

pub const BITS: usize = 64usize

Size of the inner integer in bits.

pub const BYTES: usize = 8usize

Size of the inner integer in bytes.