Trait linera_alloy::signers::k256::elliptic_curve::bigint::ArrayEncoding

pub trait ArrayEncoding: Encoding {
    type ByteSize: ArrayLength<u8> + Add + Eq + Ord + Unsigned;

    // Required methods
    fn from_be_byte_array(bytes: GenericArray<u8, Self::ByteSize>) -> Self;
    fn from_le_byte_array(bytes: GenericArray<u8, Self::ByteSize>) -> Self;
    fn to_be_byte_array(&self) -> GenericArray<u8, Self::ByteSize>;
    fn to_le_byte_array(&self) -> GenericArray<u8, Self::ByteSize>;
}

Support for encoding a big integer as a GenericArray.

Required Associated Types

type ByteSize: ArrayLength<u8> + Add + Eq + Ord + Unsigned

Size of a byte array which encodes a big integer.

Required Methods

fn from_be_byte_array(bytes: GenericArray<u8, Self::ByteSize>) -> Self

Deserialize from a big-endian byte array.

fn from_le_byte_array(bytes: GenericArray<u8, Self::ByteSize>) -> Self

Deserialize from a little-endian byte array.

fn to_be_byte_array(&self) -> GenericArray<u8, Self::ByteSize>

Serialize to a big-endian byte array.

fn to_le_byte_array(&self) -> GenericArray<u8, Self::ByteSize>

Serialize to a little-endian byte array.

Object Safety

This trait is not object safe.