Trait linera_alloy::signers::k256::elliptic_curve::ops::BatchInvert

pub trait BatchInvert<FieldElements>: Sized + Invert
where
    FieldElements: ?Sized,
{
    type Output: AsRef<[Self]>;

    // Required method
    fn batch_invert(field_elements: &FieldElements) -> CtOption<Self::Output>;
}

Perform a batched inversion on a sequence of field elements (i.e. base field elements or scalars) at an amortized cost that should be practically as efficient as a single inversion.

Required Associated Types

type Output: AsRef<[Self]>

The output of batch inversion. A container of field elements.

Required Methods

fn batch_invert(field_elements: &FieldElements) -> CtOption<Self::Output>

Invert a batch of field elements.

Object Safety

This trait is not object safe.