Struct linera_alloy::signers::k256::elliptic_curve::scalar::ScalarPrimitive

pub struct ScalarPrimitive<C>
where
    C: Curve,
{ /* private fields */ }

Generic scalar type with primitive functionality.

This type provides a baseline level of scalar arithmetic functionality which is always available for all curves, regardless of if they implement any arithmetic traits.

serde support

When the optional serde feature of this create is enabled, [Serialize] and [Deserialize] impls are provided for this type.

The serialization is a fixed-width big endian encoding. When used with textual formats, the binary data is encoded as hexadecimal.

Implementations

impl<C> ScalarPrimitive<C>
where C: Curve,

pub const ZERO: ScalarPrimitive<C> = _

Zero scalar.

pub const ONE: ScalarPrimitive<C> = _

Multiplicative identity.

pub const MODULUS: <C as Curve>::Uint = C::ORDER

Scalar modulus.

pub fn random(rng: &mut impl CryptoRngCore) -> ScalarPrimitive<C>

Generate a random ScalarPrimitive.

pub fn new(uint: <C as Curve>::Uint) -> CtOption<ScalarPrimitive<C>>

Create a new scalar from Curve::Uint.

pub fn from_bytes( bytes: &GenericArray<u8, <C as Curve>::FieldBytesSize>, ) -> CtOption<ScalarPrimitive<C>>

Decode ScalarPrimitive from a serialized field element

pub fn from_slice(slice: &[u8]) -> Result<ScalarPrimitive<C>, Error>

Decode ScalarPrimitive from a big endian byte slice.

pub fn as_uint(&self) -> &<C as Curve>::Uint

Borrow the inner C::Uint.

pub fn as_limbs(&self) -> &[Limb]

Borrow the inner limbs as a slice.

pub fn is_zero(&self) -> Choice

Is this ScalarPrimitive value equal to zero?

pub fn is_even(&self) -> Choice

Is this ScalarPrimitive value even?

pub fn is_odd(&self) -> Choice

Is this ScalarPrimitive value odd?

pub fn to_bytes(&self) -> GenericArray<u8, <C as Curve>::FieldBytesSize>

Encode ScalarPrimitive as a serialized field element.

pub fn to_uint(&self) -> <C as Curve>::Uint

Convert to a C::Uint.