Trait linera_alloy::signers::k256::elliptic_curve::sec1::ValidatePublicKey

pub trait ValidatePublicKey: Curve
where
    Self::FieldBytesSize: ModulusSize,
{
    // Provided method
    fn validate_public_key(
        secret_key: &SecretKey<Self>,
        public_key: &EncodedPoint<Self::FieldBytesSize>,
    ) -> Result<(), Error> { ... }
}

Validate that the given EncodedPoint represents the encoded public key value of the given secret.

Curve implementations which also impl CurveArithmetic will receive a blanket default impl of this trait.

Provided Methods

fn validate_public_key( secret_key: &SecretKey<Self>, public_key: &EncodedPoint<Self::FieldBytesSize>, ) -> Result<(), Error>

Validate that the given EncodedPoint is a valid public key for the provided secret value.

Object Safety

This trait is not object safe.