Struct linera_alloy::signers::k256::ProjectivePoint

pub struct ProjectivePoint { /* private fields */ }

A point on the secp256k1 curve in projective coordinates.

Implementations

impl ProjectivePoint

pub const IDENTITY: ProjectivePoint = _

Additive identity of the group: the point at infinity.

pub const GENERATOR: ProjectivePoint = _

Base point of secp256k1.

pub const fn identity() -> ProjectivePoint

๐Deprecated since 0.10.2: use ProjectivePoint::IDENTITY instead

Returns the additive identity of SECP256k1, also known as the โneutral elementโ or โpoint at infinityโ.

pub fn generator() -> ProjectivePoint

๐Deprecated since 0.10.2: use ProjectivePoint::GENERATOR instead

Returns the base point of SECP256k1.

pub fn to_affine(&self) -> AffinePoint

Returns the affine representation of this point.

pub fn double(&self) -> ProjectivePoint

Doubles this point.

pub fn endomorphism(&self) -> ProjectivePoint

Calculates SECP256k1 endomorphism: self * lambda.

pub fn eq_affine(&self, other: &AffinePoint) -> Choice

Check whether self is equal to an affine point.

This is a lot faster than first converting self to an AffinePoint and then doing the comparison. It is a little bit faster than converting other to a ProjectivePoint first.