Struct linera_alloy::signers::wallet::Wallet

pub struct Wallet<D> { /* private fields */ }

An Ethereum private-public key pair which can be used for signing messages.

Examples

Signing and Verifying a message

The wallet can be used to produce ECDSA Signature objects, which can be then verified. Note that this uses eip191_hash_message under the hood which will prefix the message being hashed with the Ethereum Signed Message domain separator.

use linera_alloy_signer::{Signer, SignerSync};

let wallet = linera_alloy_signer_wallet::LocalWallet::random();

// Optionally, the wallet's chain id can be set, in order to use EIP-155
// replay protection with different chains
let wallet = wallet.with_chain_id(Some(1337));

// The wallet can be used to sign messages
let message = b"hello";
let signature = wallet.sign_message_sync(message)?;
assert_eq!(signature.recover_address_from_msg(&message[..]).unwrap(), wallet.address());

// LocalWallet is clonable:
let wallet_clone = wallet.clone();
let signature2 = wallet_clone.sign_message_sync(message)?;
assert_eq!(signature, signature2);

Implementations

impl Wallet<SigningKey<Secp256k1>>

pub fn from_signing_key( signer: SigningKey<Secp256k1>, ) -> Wallet<SigningKey<Secp256k1>>

Creates a new Wallet instance from a SigningKey.

This can also be used to create a Wallet from a SecretKey. See also the From implementations.

pub fn from_bytes( bytes: &FixedBytes<32>, ) -> Result<Wallet<SigningKey<Secp256k1>>, Error>

Creates a new Wallet instance from a raw scalar serialized as a B256 byte array.

This is identical to from_field_bytes.

pub fn from_field_bytes( bytes: &GenericArray<u8, <Secp256k1 as Curve>::FieldBytesSize>, ) -> Result<Wallet<SigningKey<Secp256k1>>, Error>

Creates a new Wallet instance from a raw scalar serialized as a FieldBytes byte array.

pub fn from_slice(bytes: &[u8]) -> Result<Wallet<SigningKey<Secp256k1>>, Error>

Creates a new Wallet instance from a raw scalar serialized as a byte slice.

Byte slices shorter than the field size (32 bytes) are handled by zero padding the input.

pub fn random() -> Wallet<SigningKey<Secp256k1>>

Creates a new random keypair seeded with rand::thread_rng().

pub fn random_with<R>(rng: &mut R) -> Wallet<SigningKey<Secp256k1>>
where R: Rng + CryptoRng,

Creates a new random keypair seeded with the provided RNG.

pub fn as_nonzero_scalar(&self) -> &NonZeroScalar<Secp256k1>

Borrow the secret NonZeroScalar value for this key.

⚠️ Warning

This value is key material.

Please treat it with the care it deserves!

pub fn to_bytes(&self) -> FixedBytes<32>

Serialize this Wallet’s SigningKey as a B256 byte array.

pub fn to_field_bytes( &self, ) -> GenericArray<u8, <Secp256k1 as Curve>::FieldBytesSize>

Serialize this Wallet’s SigningKey as a FieldBytes byte array.

impl<D> Wallet<D>
where D: PrehashSigner<(Signature<Secp256k1>, RecoveryId)>,

pub const fn new_with_signer( signer: D, address: Address, chain_id: Option<u64>, ) -> Wallet<D>

Construct a new wallet with an external PrehashSigner.

pub const fn signer(&self) -> &D

Returns this wallet’s signer.

pub fn into_signer(self) -> D

Consumes this wallet and returns its signer.

pub const fn address(&self) -> Address

Returns this wallet’s chain ID.

pub const fn chain_id(&self) -> Option<u64>

Returns this wallet’s chain ID.