Trait linera_alloy::sol_types::SolCall

pub trait SolCall: Sized {
    type Parameters<'a>: SolType<Token<'a> = Self::Token<'a>>;
    type Token<'a>: TokenSeq<'a>;
    type Return;
    type ReturnTuple<'a>: SolType<Token<'a> = Self::ReturnToken<'a>>;
    type ReturnToken<'a>: TokenSeq<'a>;

    const SIGNATURE: &'static str;
    const SELECTOR: [u8; 4];

    // Required methods
    fn new(tuple: <Self::Parameters<'_> as SolType>::RustType) -> Self;
    fn tokenize(&self) -> Self::Token<'_>;
    fn abi_decode_returns(
        data: &[u8],
        validate: bool,
    ) -> Result<Self::Return, Error>;

    // Provided methods
    fn abi_encoded_size(&self) -> usize { ... }
    fn abi_decode_raw(data: &[u8], validate: bool) -> Result<Self, Error> { ... }
    fn abi_decode(data: &[u8], validate: bool) -> Result<Self, Error> { ... }
    fn abi_encode_raw(&self, out: &mut Vec<u8>) { ... }
    fn abi_encode(&self) -> Vec<u8>  { ... }
    fn abi_encode_returns<'a, E>(e: &'a E) -> Vec<u8> 
       where E: SolTypeValue<Self::ReturnTuple<'a>> { ... }
}

A Solidity function call.

Implementer’s Guide

It should not be necessary to implement this trait manually. Instead, use the sol! procedural macro to parse Solidity syntax into types that implement this trait.

Required Associated Types

type Parameters<'a>: SolType<Token<'a> = Self::Token<'a>>

The underlying tuple type which represents this type’s arguments.

If this type has no arguments, this will be the unit type ().

type Token<'a>: TokenSeq<'a>

The arguments’ corresponding TokenSeq type.

type Return

The function’s return struct.

type ReturnTuple<'a>: SolType<Token<'a> = Self::ReturnToken<'a>>

The underlying tuple type which represents this type’s return values.

If this type has no return values, this will be the unit type ().

type ReturnToken<'a>: TokenSeq<'a>

The returns’ corresponding TokenSeq type.

Required Associated Constants

const SIGNATURE: &'static str

The function’s ABI signature.

const SELECTOR: [u8; 4]

The function selector: keccak256(SIGNATURE)[0..4]

Required Methods

fn new(tuple: <Self::Parameters<'_> as SolType>::RustType) -> Self

Convert from the tuple type used for ABI encoding and decoding.

fn tokenize(&self) -> Self::Token<'_>

Tokenize the call’s arguments.

fn abi_decode_returns( data: &[u8], validate: bool, ) -> Result<Self::Return, Error>

ABI decode this call’s return values from the given slice.

Provided Methods

fn abi_encoded_size(&self) -> usize

The size of the encoded data in bytes, without its selector.

fn abi_decode_raw(data: &[u8], validate: bool) -> Result<Self, Error>

ABI decode this call’s arguments from the given slice, without its selector.

fn abi_decode(data: &[u8], validate: bool) -> Result<Self, Error>

ABI decode this call’s arguments from the given slice, with the selector.

fn abi_encode_raw(&self, out: &mut Vec<u8>)

ABI encode the call to the given buffer without its selector.

fn abi_encode(&self) -> Vec<u8>

ABI encode the call to the given buffer with its selector.

fn abi_encode_returns<'a, E>(e: &'a E) -> Vec<u8>
where E: SolTypeValue<Self::ReturnTuple<'a>>,

ABI encode the call’s return values.

Object Safety

This trait is not object safe.