Struct linera_alloy::transports::BoxTransport

pub struct BoxTransport { /* private fields */ }

A boxed, Clone-able Transport trait object.

This type allows RPC clients to use a type-erased transport. It is Clone and Send + Sync, and implements Transport. This allows for complex behavior abstracting across several different clients with different transport types.

Most higher-level types will be generic over T: Transport = BoxTransport. This allows paramterization with a concrete type, while hiding this complexity from the library consumer.

Implementations

impl BoxTransport

pub fn new<T>(inner: T) -> BoxTransport
where T: Transport + Clone + Send + Sync,

Instantiate a new box transport from a suitable transport.

pub fn as_any(&self) -> &(dyn Any + 'static)

Returns a reference to the inner transport.